Important fix.
Important bug fixed:
Error on selecting FK entities. The build projection method was just
taking the first mapping that matches the property type. The problem was
the parent entity has two propertis of the same type so the second one
was never added in the projection. Now the parent path was used to
filter the mapping, fixing the bug.

Issues: https://github.com/marionzr/Nzr.Orm/issues/35