Fixed a bug
Where using Where filter with IN or NOT IN operators and
the args is empty and exception was throw. Now, a new option
was introduced, enabled by default, to handle this problem
by changing the Where filter from 'WHERE column IN ()' or
'WHERE column NOT IN ()' to WHERE 'column IN ()' = 'empty args'
and 'WHERE column NOT IN ()' <> 'empty args'.

A new test coverage report was also included. The nexts releases
will try to reach a 96% of line and branch test coverage.

Issues: https://github.com/marionzr/Nzr.Orm/issues/36